+++ src/test/java/subway/service/LineServiceTest.java
@@ -16,8 +16,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import subway.controller.exception.BusinessException;
 import subway.domain.Line;
 import subway.domain.Subway;
-import subway.dto.LineDto;
-import subway.dto.SectionDto;
+import subway.service.dto.LineDto;
+import subway.service.dto.SectionDto;
 import subway.dto.response.LineResponse;
 import subway.repository.LineRepository;
 import subway.repository.SubwayRepository;
